pyproject.toml
[project]
name = "medwise-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "chromadb>=1.4.0",
    "dagshub>=0.6.4",
    "dvc>=3.66.1",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "langchain>=1.2.3",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-huggingface>=1.2.0",
    "langchain-ollama>=1.0.1",
    "mlflow>=3.8.1",
    "numpy>=2.4.1",
    "pydantic>=2.12.5",
    "pypdf>=6.6.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.2.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.5",
    "uvicorn>=0.40.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["src/tests"]
